refactor(auth): split large files and improve code organization
Auth High & Medium Priority Tasks completed:

Task 1: AuthContext.tsx split into 3 files
- authMapper.ts: Supabase-to-App type mapping
- authActions.ts: All auth operations (login, register, etc.)
- AuthContext.tsx: Reduced to state + effects only

Task 2: Password Reset UI
- SetPasswordScreen.tsx with proper routing

Task 3-4: Screen styles extracted
- LoginScreen.styles.ts + useLoginForm.ts
- RegisterScreen.styles.ts + useRegisterForm.ts

Task 5: Hardcoded colors replaced
- warningMedium, errorSubtle, shadowModal tokens

Task 6: Close buttons 44px touch targets

Task 7: console.log → DEV guards (import.meta.env.DEV)

Task 8: authService.ts deprecated
- getUserById moved to authHelpers.ts

Task 9: Focus-visible styles added
- global.css + Button.tsx
