+++ src/features/auth/components/GuestBanner.tsx
@@ -165,8 +165,10 @@ const styles: Record<string, CSSProperties> = {
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
-    width: '32px',
-    height: '32px',
+    width: '44px',
+    height: '44px',
+    minWidth: '44px',
+    minHeight: '44px',
     padding: 0,
     border: 'none',
     background: 'transparent',
