+++ src/features/auth/components/InviteDialog.tsx
@@ -170,7 +170,7 @@ const styles: Record<string, CSSProperties> = {
     left: 0,
     right: 0,
     bottom: 0,
-    background: 'rgba(0, 0, 0, 0.5)',
+    background: cssVars.colors.overlay,
     display: 'flex',
     justifyContent: 'center',
     alignItems: 'center',
