+++ src/features/auth/components/LoginScreen.styles.ts
@@ -1,0 +1,293 @@
+/**
+ * LoginScreen Styles
+ *
+ * Extracted from LoginScreen.tsx for better maintainability.
+ * All styles use design tokens for consistency.
+ *
+ * @see docs/concepts/ANMELDUNG-KONZEPT.md
+ */
+
+import type { CSSProperties } from 'react';
+import { cssVars } from '../../../design-tokens';
+
+export const loginStyles: Record<string, CSSProperties> = {
+  backdrop: {
+    position: 'fixed',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: cssVars.spacing.lg,
+    background: cssVars.colors.backdrop,
+    backdropFilter: 'blur(4px)',
+    WebkitBackdropFilter: 'blur(4px)',
+    zIndex: 1000,
+  },
+  card: {
+    position: 'relative',
+    width: '100%',
+    maxWidth: '400px',
+    maxHeight: '90vh',
+    overflowY: 'auto',
+    padding: cssVars.spacing.xl,
+    background: cssVars.colors.surfaceSolid,
+    borderRadius: cssVars.borderRadius.lg,
+    border: `1px solid ${cssVars.colors.border}`,
+    boxShadow: `0 8px 32px ${cssVars.colors.shadowModal}`,
+  },
+  header: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    marginBottom: cssVars.spacing.md,
+    minHeight: '44px',
+    gap: cssVars.spacing.sm,
+    flexWrap: 'nowrap' as const,
+  },
+  backButton: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: cssVars.spacing.xs,
+    padding: `${cssVars.spacing.xs} ${cssVars.spacing.sm}`,
+    background: 'transparent',
+    border: 'none',
+    borderRadius: cssVars.borderRadius.sm,
+    color: cssVars.colors.textSecondary,
+    fontSize: cssVars.fontSizes.sm,
+    cursor: 'pointer',
+    transition: 'color 0.2s, background 0.2s',
+    whiteSpace: 'nowrap' as const,
+    flexShrink: 0,
+    minHeight: '44px',
+  },
+  closeButton: {
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    width: '44px',
+    height: '44px',
+    minWidth: '44px',
+    minHeight: '44px',
+    marginLeft: 'auto',
+    background: 'transparent',
+    border: 'none',
+    borderRadius: cssVars.borderRadius.sm,
+    color: cssVars.colors.textSecondary,
+    fontSize: cssVars.fontSizes.lg,
+    cursor: 'pointer',
+    transition: 'color 0.2s, background 0.2s',
+  },
+  title: {
+    fontSize: cssVars.fontSizes.xxl,
+    fontWeight: cssVars.fontWeights.bold,
+    color: cssVars.colors.textPrimary,
+    margin: 0,
+    marginBottom: cssVars.spacing.sm,
+    textAlign: 'center',
+  },
+  subtitle: {
+    fontSize: cssVars.fontSizes.md,
+    color: cssVars.colors.textSecondary,
+    margin: 0,
+    marginBottom: cssVars.spacing.lg,
+    textAlign: 'center',
+  },
+  form: {
+    display: 'flex',
+    flexDirection: 'column',
+    gap: cssVars.spacing.md,
+  },
+  inputGroup: {
+    display: 'flex',
+    flexDirection: 'column',
+    gap: cssVars.spacing.xs,
+  },
+  label: {
+    fontSize: cssVars.fontSizes.sm,
+    fontWeight: cssVars.fontWeights.medium,
+    color: cssVars.colors.textSecondary,
+  },
+  input: {
+    width: '100%',
+    height: '48px',
+    padding: `0 ${cssVars.spacing.md}`,
+    fontSize: cssVars.fontSizes.lg,
+    color: cssVars.colors.textPrimary,
+    background: cssVars.colors.surfaceSolid,
+    border: `1px solid ${cssVars.colors.border}`,
+    borderRadius: cssVars.borderRadius.md,
+    boxSizing: 'border-box',
+    transition: 'border-color 0.2s ease',
+  },
+  inputError: {
+    borderColor: cssVars.colors.error,
+  },
+  forgotPasswordLink: {
+    alignSelf: 'flex-end',
+    marginTop: cssVars.spacing.xs,
+    padding: 0,
+    background: 'transparent',
+    border: 'none',
+    color: cssVars.colors.textMuted,
+    fontSize: cssVars.fontSizes.sm,
+    cursor: 'pointer',
+    textDecoration: 'underline',
+  },
+  errorText: {
+    fontSize: cssVars.fontSizes.sm,
+    color: cssVars.colors.error,
+  },
+  button: {
+    minHeight: '56px',
+  },
+  modeToggle: {
+    display: 'block',
+    width: '100%',
+    marginTop: cssVars.spacing.sm,
+    padding: cssVars.spacing.sm,
+    background: 'transparent',
+    border: 'none',
+    color: cssVars.colors.textMuted,
+    fontSize: cssVars.fontSizes.sm,
+    cursor: 'pointer',
+    textAlign: 'center',
+    textDecoration: 'underline',
+  },
+  divider: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: cssVars.spacing.md,
+    margin: `${cssVars.spacing.lg} 0`,
+  },
+  dividerLine: {
+    flex: 1,
+    height: '1px',
+    background: cssVars.colors.border,
+  },
+  dividerText: {
+    fontSize: cssVars.fontSizes.sm,
+    color: cssVars.colors.textSecondary,
+  },
+  googleButton: {
+    minHeight: '48px',
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    gap: cssVars.spacing.sm,
+  },
+  googleIcon: {
+    fontWeight: cssVars.fontWeights.bold,
+    fontSize: cssVars.fontSizes.lg,
+    color: cssVars.colors.primary,
+  },
+  ghostButton: {
+    minHeight: '48px',
+    marginTop: cssVars.spacing.sm,
+  },
+  footer: {
+    marginTop: cssVars.spacing.lg,
+    fontSize: cssVars.fontSizes.sm,
+    color: cssVars.colors.textSecondary,
+    textAlign: 'center',
+  },
+  link: {
+    color: cssVars.colors.primary,
+    background: 'none',
+    border: 'none',
+    padding: 0,
+    font: 'inherit',
+    cursor: 'pointer',
+    fontWeight: cssVars.fontWeights.medium,
+    textDecoration: 'none',
+  },
+  successIcon: {
+    width: '64px',
+    height: '64px',
+    margin: '0 auto',
+    marginBottom: cssVars.spacing.md,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    fontSize: cssVars.fontSizes.xxl,
+    color: cssVars.colors.background,
+    background: cssVars.colors.primary,
+    borderRadius: cssVars.borderRadius.full,
+  },
+  successTitle: {
+    fontSize: cssVars.fontSizes.xl,
+    fontWeight: cssVars.fontWeights.bold,
+    color: cssVars.colors.textPrimary,
+    margin: 0,
+    marginBottom: cssVars.spacing.sm,
+    textAlign: 'center',
+  },
+  successText: {
+    fontSize: cssVars.fontSizes.md,
+    color: cssVars.colors.textSecondary,
+    margin: 0,
+    textAlign: 'center',
+    lineHeight: '1.5',
+  },
+  successCloseButton: {
+    position: 'absolute',
+    top: cssVars.spacing.md,
+    right: cssVars.spacing.md,
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'center',
+    width: '44px',
+    height: '44px',
+    minWidth: '44px',
+    minHeight: '44px',
+    background: 'transparent',
+    border: 'none',
+    borderRadius: cssVars.borderRadius.sm,
+    color: cssVars.colors.textSecondary,
+    fontSize: cssVars.fontSizes.lg,
+    cursor: 'pointer',
+    transition: 'color 0.2s, background 0.2s',
+  },
+  offlineBanner: {
+    display: 'flex',
+    alignItems: 'center',
+    gap: cssVars.spacing.sm,
+    padding: cssVars.spacing.md,
+    marginBottom: cssVars.spacing.md,
+    backgroundColor: cssVars.colors.warningMedium,
+    border: `1px solid ${cssVars.colors.warning}`,
+    borderRadius: cssVars.borderRadius.md,
+  },
+  offlineIcon: {
+    fontSize: cssVars.fontSizes.lg,
+    flexShrink: 0,
+  },
+  offlineTextContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    flex: 1,
+    gap: '2px',
+  },
+  offlineTitle: {
+    fontSize: cssVars.fontSizes.sm,
+    fontWeight: cssVars.fontWeights.semibold,
+    color: cssVars.colors.warning,
+  },
+  offlineSubtitle: {
+    fontSize: cssVars.fontSizes.xs,
+    color: cssVars.colors.textSecondary,
+  },
+  offlineRetryButton: {
+    padding: `${cssVars.spacing.xs} ${cssVars.spacing.sm}`,
+    background: 'transparent',
+    border: `1px solid ${cssVars.colors.warning}`,
+    borderRadius: cssVars.borderRadius.sm,
+    color: cssVars.colors.warning,
+    fontSize: cssVars.fontSizes.sm,
+    fontWeight: cssVars.fontWeights.medium,
+    cursor: 'pointer',
+    flexShrink: 0,
+  },
+};
