+++ src/features/auth/components/RegisterScreen.tsx
@@ -14,137 +14,13 @@
  * @see docs/concepts/ANMELDUNG-KONZEPT.md
  */
 
-import React, { useState, useEffect, CSSProperties } from 'react';
+import React, { useState, useEffect } from 'react';
 import { cssVars } from '../../../design-tokens';
 import { Button } from '../../../components/ui/Button';
 import { useToast } from '../../../components/ui/Toast';
 import { useAuth } from '../hooks/useAuth';
-
-// =============================================================================
-// EMAIL VALIDATION
-// =============================================================================
-
-/** Common valid TLDs */
-const VALID_TLDS = new Set([
-  // Generic
-  'com', 'net', 'org', 'edu', 'gov', 'info', 'biz', 'io', 'co', 'app', 'dev',
-  // European
-  'de', 'at', 'ch', 'nl', 'be', 'fr', 'uk', 'es', 'it', 'pt', 'pl', 'cz', 'eu',
-  // Other major
-  'us', 'ca', 'au', 'nz', 'jp', 'kr', 'cn', 'in', 'br', 'mx', 'ru',
-]);
-
-/** Common disposable email domains to block */
-const DISPOSABLE_DOMAINS = new Set([
-  'tempmail.com', 'throwaway.email', '10minutemail.com', 'guerrillamail.com',
-  'mailinator.com', 'trashmail.com', 'fakeinbox.com', 'temp-mail.org',
-  'disposablemail.com', 'yopmail.com', 'tempail.com', 'mohmal.com',
-  'getnada.com', 'mailnesia.com', 'emailondeck.com', 'dispostable.com',
-]);
-
-/** Common email provider typos */
-const EMAIL_TYPO_SUGGESTIONS: Record<string, string> = {
-  'gmial.com': 'gmail.com',
-  'gmal.com': 'gmail.com',
-  'gamil.com': 'gmail.com',
-  'gnail.com': 'gmail.com',
-  'gmail.de': 'gmail.com', // Gmail doesn't have .de
-  'gmx.com': 'gmx.de', // GMX is mostly .de
-  'outook.com': 'outlook.com',
-  'outlok.com': 'outlook.com',
-  'hotmal.com': 'hotmail.com',
-  'hotmai.com': 'hotmail.com',
-  'yahooo.com': 'yahoo.com',
-  'yaho.com': 'yahoo.com',
-  'web.com': 'web.de',
-};
-
-interface EmailValidationResult {
-  isValid: boolean;
-  error?: string;
-  suggestion?: string;
-}
-
-/**
- * Enhanced email validation
- * Checks format, TLD validity, disposable domains, and common typos
- */
-function validateEmail(email: string): EmailValidationResult {
-  const trimmed = email.trim().toLowerCase();
-
-  // Basic format check
-  if (!trimmed) {
-    return { isValid: false, error: 'E-Mail ist erforderlich' };
-  }
-
-  // More comprehensive regex: local@domain.tld
-  const emailRegex = /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*\.[a-zA-Z]{2,}$/;
-  if (!emailRegex.test(trimmed)) {
-    return { isValid: false, error: 'Bitte gib eine gültige E-Mail-Adresse ein' };
-  }
-
-  // Extract domain parts
-  const parts = trimmed.split('@');
-  if (parts.length !== 2) {
-    return { isValid: false, error: 'Ungültiges E-Mail-Format' };
-  }
-
-  const [localPart, domain] = parts;
-
-  // Local part checks
-  if (localPart.length < 1) {
-    return { isValid: false, error: 'E-Mail-Adresse vor @ ist zu kurz' };
-  }
-  if (localPart.length > 64) {
-    return { isValid: false, error: 'E-Mail-Adresse vor @ ist zu lang' };
-  }
-
-  // Domain checks
-  if (domain.length < 4) {
-    return { isValid: false, error: 'Domain ist zu kurz' };
-  }
-  if (domain.length > 255) {
-    return { isValid: false, error: 'Domain ist zu lang' };
-  }
-
-  // Check for typos in common providers
-  const suggestion = EMAIL_TYPO_SUGGESTIONS[domain];
-  if (suggestion) {
-    return {
-      isValid: false,
-      error: `Meintest du ${localPart}@${suggestion}?`,
-      suggestion: `${localPart}@${suggestion}`,
-    };
-  }
-
-  // Check for disposable email domains
-  if (DISPOSABLE_DOMAINS.has(domain)) {
-    return {
-      isValid: false,
-      error: 'Wegwerf-E-Mail-Adressen sind nicht erlaubt',
-    };
-  }
-
-  // Extract and validate TLD
-  const domainParts = domain.split('.');
-  const tld = domainParts[domainParts.length - 1];
-
-  if (tld.length < 2) {
-    return { isValid: false, error: 'Ungültige Domain-Endung' };
-  }
-
-  // Check if TLD is in our known list (warning, not blocking)
-  // We don't block unknown TLDs as new ones are created regularly
-  if (!VALID_TLDS.has(tld) && tld.length > 6) {
-    // Very long unknown TLDs are suspicious
-    return {
-      isValid: false,
-      error: 'Bitte überprüfe die Domain-Endung',
-    };
-  }
-
-  return { isValid: true };
-}
+import { validateEmail } from '../utils/emailValidation';
+import { registerStyles as styles } from './RegisterScreen.styles';
 
 interface RegisterScreenProps {
   /** Called when registration is successful */
@@ -628,285 +504,6 @@ export const RegisterScreen: React.FC<RegisterScreenProps> = ({
       </div>
     </div>
   );
-};
-
-// Styles using design tokens
-const styles: Record<string, CSSProperties> = {
-  backdrop: {
-    position: 'fixed',
-    top: 0,
-    left: 0,
-    right: 0,
-    bottom: 0,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: cssVars.spacing.lg,
-    background: cssVars.colors.backdrop,
-    backdropFilter: 'blur(4px)',
-    WebkitBackdropFilter: 'blur(4px)',
-    zIndex: 1000,
-  },
-  card: {
-    position: 'relative',
-    width: '100%',
-    maxWidth: '400px',
-    maxHeight: '90vh',
-    overflowY: 'auto',
-    padding: cssVars.spacing.xl,
-    background: cssVars.colors.surfaceSolid,
-    borderRadius: cssVars.borderRadius.lg,
-    border: `1px solid ${cssVars.colors.border}`,
-    boxShadow: '0 8px 32px rgba(0, 0, 0, 0.4)',
-  },
-  header: {
-    display: 'flex',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: cssVars.spacing.md,
-    minHeight: '32px',
-  },
-  backButton: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: cssVars.spacing.xs,
-    padding: `${cssVars.spacing.xs} ${cssVars.spacing.sm}`,
-    background: 'transparent',
-    border: 'none',
-    borderRadius: cssVars.borderRadius.sm,
-    color: cssVars.colors.textSecondary,
-    fontSize: cssVars.fontSizes.sm,
-    cursor: 'pointer',
-    transition: 'color 0.2s, background 0.2s',
-  },
-  closeButton: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '32px',
-    height: '32px',
-    marginLeft: 'auto',
-    background: 'transparent',
-    border: 'none',
-    borderRadius: cssVars.borderRadius.sm,
-    color: cssVars.colors.textSecondary,
-    fontSize: cssVars.fontSizes.lg,
-    cursor: 'pointer',
-    transition: 'color 0.2s, background 0.2s',
-  },
-  title: {
-    fontSize: cssVars.fontSizes.xxl,
-    fontWeight: cssVars.fontWeights.bold,
-    color: cssVars.colors.textPrimary,
-    margin: 0,
-    marginBottom: cssVars.spacing.sm,
-    textAlign: 'center',
-  },
-  subtitle: {
-    fontSize: cssVars.fontSizes.md,
-    color: cssVars.colors.textSecondary,
-    margin: 0,
-    marginBottom: cssVars.spacing.lg,
-    textAlign: 'center',
-  },
-  form: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: cssVars.spacing.md,
-  },
-  inputGroup: {
-    display: 'flex',
-    flexDirection: 'column',
-    gap: cssVars.spacing.xs,
-  },
-  label: {
-    fontSize: cssVars.fontSizes.sm,
-    fontWeight: cssVars.fontWeights.medium,
-    color: cssVars.colors.textSecondary,
-  },
-  input: {
-    width: '100%',
-    height: '48px',
-    padding: `0 ${cssVars.spacing.md}`,
-    fontSize: cssVars.fontSizes.lg,
-    color: cssVars.colors.textPrimary,
-    background: cssVars.colors.surfaceSolid,
-    border: `1px solid ${cssVars.colors.border}`,
-    borderRadius: cssVars.borderRadius.md,
-    boxSizing: 'border-box',
-    transition: 'border-color 0.2s ease',
-  },
-  inputError: {
-    borderColor: cssVars.colors.error,
-  },
-  errorText: {
-    fontSize: cssVars.fontSizes.sm,
-    color: cssVars.colors.error,
-  },
-  hint: {
-    fontSize: cssVars.fontSizes.xs,
-    color: cssVars.colors.textMuted,
-    marginTop: cssVars.spacing.xs,
-  },
-  suggestionButton: {
-    marginTop: cssVars.spacing.xs,
-    padding: `${cssVars.spacing.xs} ${cssVars.spacing.sm}`,
-    fontSize: cssVars.fontSizes.sm,
-    color: cssVars.colors.primary,
-    background: cssVars.colors.primarySubtle,
-    border: `1px solid ${cssVars.colors.primary}`,
-    borderRadius: cssVars.borderRadius.sm,
-    cursor: 'pointer',
-    fontWeight: cssVars.fontWeights.medium,
-  },
-  generalError: {
-    padding: cssVars.spacing.md,
-    background: 'rgba(239, 68, 68, 0.1)',
-    border: `1px solid ${cssVars.colors.error}`,
-    borderRadius: cssVars.borderRadius.md,
-    color: cssVars.colors.error,
-    fontSize: cssVars.fontSizes.sm,
-    textAlign: 'center',
-  },
-  button: {
-    minHeight: '56px',
-    marginTop: cssVars.spacing.sm,
-  },
-  divider: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: cssVars.spacing.md,
-    margin: `${cssVars.spacing.lg} 0`,
-  },
-  dividerLine: {
-    flex: 1,
-    height: '1px',
-    background: cssVars.colors.border,
-  },
-  dividerText: {
-    fontSize: cssVars.fontSizes.sm,
-    color: cssVars.colors.textSecondary,
-  },
-  googleButton: {
-    minHeight: '48px',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    gap: cssVars.spacing.sm,
-  },
-  googleIcon: {
-    fontWeight: cssVars.fontWeights.bold,
-    fontSize: cssVars.fontSizes.lg,
-    color: cssVars.colors.primary,
-  },
-  footer: {
-    marginTop: cssVars.spacing.lg,
-    fontSize: cssVars.fontSizes.sm,
-    color: cssVars.colors.textSecondary,
-    textAlign: 'center',
-  },
-  link: {
-    color: cssVars.colors.primary,
-    background: 'none',
-    border: 'none',
-    padding: 0,
-    font: 'inherit',
-    cursor: 'pointer',
-    fontWeight: cssVars.fontWeights.medium,
-    textDecoration: 'none',
-  },
-  legal: {
-    marginTop: cssVars.spacing.md,
-    fontSize: cssVars.fontSizes.xs,
-    color: cssVars.colors.textMuted,
-    textAlign: 'center',
-    lineHeight: '1.5',
-  },
-  successIcon: {
-    width: '64px',
-    height: '64px',
-    margin: '0 auto',
-    marginBottom: cssVars.spacing.md,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    fontSize: cssVars.fontSizes.xxl,
-    color: cssVars.colors.background,
-    background: cssVars.colors.primary,
-    borderRadius: cssVars.borderRadius.full,
-  },
-  successTitle: {
-    fontSize: cssVars.fontSizes.xl,
-    fontWeight: cssVars.fontWeights.bold,
-    color: cssVars.colors.textPrimary,
-    margin: 0,
-    marginBottom: cssVars.spacing.sm,
-    textAlign: 'center',
-  },
-  successText: {
-    fontSize: cssVars.fontSizes.md,
-    color: cssVars.colors.textSecondary,
-    margin: 0,
-    textAlign: 'center',
-    lineHeight: '1.5',
-  },
-  successCloseButton: {
-    position: 'absolute',
-    top: cssVars.spacing.md,
-    right: cssVars.spacing.md,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '32px',
-    height: '32px',
-    background: 'transparent',
-    border: 'none',
-    borderRadius: cssVars.borderRadius.sm,
-    color: cssVars.colors.textSecondary,
-    fontSize: cssVars.fontSizes.lg,
-    cursor: 'pointer',
-    transition: 'color 0.2s, background 0.2s',
-  },
-  offlineBanner: {
-    display: 'flex',
-    alignItems: 'center',
-    gap: cssVars.spacing.sm,
-    padding: cssVars.spacing.md,
-    marginBottom: cssVars.spacing.md,
-    backgroundColor: 'rgba(245, 158, 11, 0.1)',
-    border: `1px solid ${cssVars.colors.warning}`,
-    borderRadius: cssVars.borderRadius.md,
-  },
-  offlineIcon: {
-    fontSize: cssVars.fontSizes.lg,
-    flexShrink: 0,
-  },
-  offlineTextContainer: {
-    display: 'flex',
-    flexDirection: 'column',
-    flex: 1,
-    gap: '2px',
-  },
-  offlineTitle: {
-    fontSize: cssVars.fontSizes.sm,
-    fontWeight: cssVars.fontWeights.semibold,
-    color: cssVars.colors.warning,
-  },
-  offlineSubtitle: {
-    fontSize: cssVars.fontSizes.xs,
-    color: cssVars.colors.textSecondary,
-  },
-  offlineRetryButton: {
-    padding: `${cssVars.spacing.xs} ${cssVars.spacing.sm}`,
-    background: 'transparent',
-    border: `1px solid ${cssVars.colors.warning}`,
-    borderRadius: cssVars.borderRadius.sm,
-    color: cssVars.colors.warning,
-    fontSize: cssVars.fontSizes.sm,
-    fontWeight: cssVars.fontWeights.medium,
-    cursor: 'pointer',
-    flexShrink: 0,
-  },
 };
 
 export default RegisterScreen;
