+++ src/features/auth/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './tokenGenerator';
 export * from './permissions';
 export * from './authHelpers';
+export * from './emailValidation';
