+++ src/styles/global.css
@@ -235,6 +235,7 @@
      =========================================== */
   --color-shadow-soft: rgba(0, 0, 0, 0.2);
   --color-shadow-medium: rgba(0, 0, 0, 0.3);
+  --color-shadow-modal: rgba(0, 0, 0, 0.4);
 
   /* ===========================================
      Colors - Danger
@@ -738,6 +739,7 @@ html[data-font-size="x-large"] {
      =========================================== */
   --color-shadow-soft: rgba(0, 0, 0, 0.08);
   --color-shadow-medium: rgba(0, 0, 0, 0.15);
+  --color-shadow-modal: rgba(0, 0, 0, 0.25);
   --shadow-sm: 0 1px 3px rgba(0, 0, 0, 0.08), 0 1px 2px rgba(0, 0, 0, 0.06);
   --shadow-md: 0 4px 6px rgba(0, 0, 0, 0.1), 0 2px 4px rgba(0, 0, 0, 0.06);
   --shadow-lg: 0 10px 15px rgba(0, 0, 0, 0.1), 0 4px 6px rgba(0, 0, 0, 0.05);
